requirements.txt
pytesseract~=0.3.10
numpy~=1.26.4
pillow~=10.2.0
flask~=3.0.2
opencv-python-headless